remove unused 이벤트 remove sign of 잔여시간 in PrepareData.vue remove unused 버튼 (every 뷰 파일) remove unnecessary 스테이터스 코드 프로세스 종료 시 다음 화면으로 넘어갈 수 있도록 router에 추가 div 상태 computed에서 관리할 수 있도록 변경(in Restartservice.vue) add polling for restart server status 잔여 시간 대신 현재 진행 중인 epoch 숫자 표시 서버와 기능 연동 add apis in api.service.js add actions in current module add check file exists in file.js add verify, translation training/state 의 리턴 data 서버에서 조립 후 response 하도록 변경 add servers restart rest and flask server and check status

+++ server/helpers/file.js
@@ -80,36 +80,42 @@ const minLineCountInDirWithPattern = async (pattern, dir) => {
 };
 
 const getLastModifiedFileNameInDir = (dir) => {
-  const files = fs.readdirSync(dir);
-  let time = 0;
   let latest = '';
-  files.forEach((file) => {
-    if (file.indexOf('.DS_') === 0 || file.indexOf('Thumbs.db') === 0) return;
-    const date = fs.statSync(`${dir}/${file}`).mtime;
-    if (time < date.getTime()) {
-      time = date.getTime();
-      latest = file;
-    }
-  });
+  if (fs.existsSync(dir)) {
+    const files = fs.readdirSync(dir);
+    let time = 0;
+    files.forEach((file) => {
+      if (file.indexOf('.DS_') === 0 || file.indexOf('Thumbs.db') === 0) return;
+      const date = fs.statSync(`${dir}/${file}`).mtime;
+      if (time < date.getTime()) {
+        time = date.getTime();
+        latest = file;
+      }
+    });
+  }
   return latest;
 };
 
 const removeAllFilesInDirSync = (path) => {
-  const files = fs.readdirSync(path);
-  files.forEach((file) => {
-    fs.unlinkSync(`${path}/${file}`);
-  });
+  if (fs.existsSync(path)) {
+    const files = fs.readdirSync(path);
+    files.forEach((file) => {
+      fs.unlinkSync(`${path}/${file}`);
+    });
+  }
 };
 
-const getTrainingSequence = (TRAINING_DIR) => {
-  const files = fs.readdirSync(TRAINING_DIR);
+const getTrainingSequence = (dir) => {
   let sequence = '';
-  const reg = new RegExp(/DATA_(EN|KO)_[\d]{8}_[\d]{6}_TRAIN.txt/g);
-  files.forEach((fileName) => {
-    if (reg.test(fileName)) {
-      sequence = fileName.match(/[\d]{8}_[\d]{6}/g);
-    }
-  });
+  if (fs.existsSync(dir)) {
+    const files = fs.readdirSync(dir);
+    const reg = new RegExp(/DATA_(EN|KO)_[\d]{8}_[\d]{6}_TRAIN.txt/g);
+    files.forEach((fileName) => {
+      if (reg.test(fileName)) {
+        sequence = fileName.match(/[\d]{8}_[\d]{6}/g);
+      }
+    });
+  }
   return sequence;
 };
